Code refactoring.
*  Add ExampleFrameListener::doRaySceneQuery() to avoid code duplication.
*  Cleanup code to in parseCommand() and fix trailing space bug.
*  Remove some redundant code in terminal command 'nick'.
*  Fix a couple of compiler warnings.
*  More snprintf() conversions.
*  A few tweaks for the windows port.
*  Remove some old debug code.

+++ src/Functions.cpp
@@ -315,7 +315,7 @@ string colourizeMaterial(string materialName, int colour)
 							uint8 *alpha = pixelData++;
 
 							int deltaR = *red - 235, deltaG = *green - 20, deltaB = *blue - 235;
-							int totalDistance = fabs(deltaR) + fabs(deltaG) + fabs(deltaB);
+							int totalDistance = fabs((double)deltaR) + fabs((double)deltaG) + fabs((double)deltaB);
 							int factor = 10;
 							// Check to see if the current pixel matches the target colour
 							if(totalDistance <= 65)
